Use proper nesting of overlapping matches by reshuffling the pairs
If two patterns match the same text, we want proper nesting, e.g. 1<2<foo>2>1 instead of 1<2<foo>1>2. To achieve that, the start patterns must be sorted ascending, but the end patterns need to be descending.
Note: We could use ingo#list#Zip() with a reverse() of the second List, but that one would only handle 20 pairs, as each one would be a separate function argument.