BUG: If a mark consists of multiple branches and one branch match is contained in another, only the contained match gets the suffix
The containing text gets a prefix, but no suffix.
,----[ example ]----
| :Mark /my planet earth\|planet/
| :MarkMarkup on "for my planet earth here" produces
| "for 1<my 1<planet>1 earth here" (bad) instead of
| "for 1<my 1<planet>1 earth>1 here" (expected)
`----
Thanks to Mun Johl for reporting this!

When substituting prefixes and suffixes with the combined pattern, each start matches /\ze\%(my planet earth\|planet\)/ (as the global flag locates all matches), but the end /\%(my planet earth\|planet\)\zs/ only matches after "planet" (as matching wants to find a (here: zero-width) match as soon as possible, but at that position the first branch match then already is out of scope).
Fix this by splitting the toplevel branches of each mark pattern into individual marks (with duplicated mark number and name), so that each is matched on its own, and both prefix and suffix are added.