:MarkMarkup: Don't clobber the search history with the last mark's pattern
PatternsOnText#Transactional#ExprEach#TransactionalSubstitute() now has an additional a:mods argument where we can pass in :keeppatterns if that's supported. As a compatibility fallback, manually remove the last search pattern from the history. As we do this only in the success case, we know that a substitution has happened and don't need a check here.